requirements.txt
asgiref==3.4.1
bleach==4.1.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
defusedxml==0.6.0
Django==3.2.3
django-allauth==0.42.0
django-authtools==1.7.0
django-ckeditor==6.1.0
django-cleanup==5.2.0
django-crispy-forms==1.9.2
django-debug-toolbar==2.2
django-environ==0.8.1
django-filter==21.1
django-js-asset==1.2.2
django-messages==0.6.0
django-multiupload==0.6.1
django-taggit==1.5.0
django-utils-six==2.0
gunicorn==20.0.4
h11==0.9.0
httptools==0.1.1
idna==2.10
isort==5.10.1
oauthlib==3.1.0
packaging==21.3
pyparsing==3.0.6
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.3.1
urllib3==1.25.9
uvicorn==0.11.5
uvloop==0.14.0
webencodings==0.5.1
websockets==8.1
whitenoise==5.1.0
